LOGIC-192 enabled specifying name for fnc
fnc enables the creation of anonymous constraints, in a similar vein
to fn in clojure core. fn has the ability to name the fn, which is
useful for traces and debugging. This codes provides a name for the
fnc, which is not only useful for debugging but also whenever
non-grounded constraints are printed.

Code is similar to what fn does in core.